ai/feedback/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# API Client & Network
openai>=1.0.0
httpx>=0.24.0
requests>=2.31.0
websockets>=12.0

# Environment & Data
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
prometheus-fastapi-instrumentator>=7.0.0
